changes to folders structure

+++ printf/ft_printf.h
@@ -18,8 +18,8 @@
 
 int	ft_printf(const char *s, ...);
 int	ft_printf_char(char c);
-int	ft_printf_hexlow(unsigned long long num);
-int	ft_printf_hexupp(unsigned long long num);
+int	ft_printf_hexlow(unsigned long num);
+int	ft_printf_hexupp(unsigned long num);
 int	ft_printf_nbr(int n);
 int	ft_printf_string(char *s);
 int	ft_printf_unbr(unsigned int nb);
